Fix redirect after going into offline mode

+++ src/services/tours/tours-service.ts
@@ -3,6 +3,7 @@ import { Tour, UpdateOrCreateTour } from '../../types/tour'
 import { ArrayApiResponse, SingleApiResponse } from '../../types/api'
 import LocalDatabaseService from '../local-database-service'
 import { TourStatusType } from '../../enums/tour-status-type'
+import { isOfflineResultMessage } from '../../utils/offline-helper'
 
 export default class ToursService extends APIService {
   private prefix: string = 'tours'
@@ -38,7 +39,7 @@ export default class ToursService extends APIService {
       this.getRequestBody('POST', tour),
       Tour
     )
-    return this.handleTourAddResult(result)
+    return this.handleTourAddResult(result, tour)
   }
 
   public async update (id: string | undefined, tour: UpdateOrCreateTour): Promise<SingleApiResponse<void>> {
@@ -59,35 +60,31 @@ export default class ToursService extends APIService {
 
   private async handleTourDeleteResult (result: SingleApiResponse<void>, id: string): Promise<SingleApiResponse<void>> {
     const localTour = await this.localDatabaseService.getOne(id)
-    if (!ToursService.isOffline(result.statusCode, result.statusMessage) && localTour) {
+    if (!isOfflineResultMessage(result.statusCode, result.statusMessage) && localTour) {
       await this.localDatabaseService.deleteTour(localTour.id)
     }
     return result
   }
 
   private async handleTourUpdateResult (result: SingleApiResponse<void>, tour: UpdateOrCreateTour, id: string | undefined): Promise<SingleApiResponse<void>> {
-    if (!ToursService.isOffline(result.statusCode, result.statusMessage)) {
+    if (!isOfflineResultMessage(result.statusCode, result.statusMessage)) {
       await this.localDatabaseService.update(id, tour, TourStatusType.SYNCED)
     }
     return result
   }
 
   private async handleGetOneResult (result: SingleApiResponse<Tour>, tourId: string | undefined): Promise<SingleApiResponse<Tour>> {
-    if (ToursService.isOffline(result.statusCode, result.statusMessage)) {
+    if (isOfflineResultMessage(result.statusCode, result.statusMessage)) {
       return result
     }
 
     return this.handleGetTour(result, tourId)
   }
 
-  private static isOffline (statusCode: number, statusMessage: string): boolean {
-    return statusCode === 500 && statusMessage === 'Failed to fetch'
-  }
-
   private async handleGetTour (result: SingleApiResponse<Tour>, tourId: string| undefined): Promise<SingleApiResponse<Tour>> {
     const localTour = await this.localDatabaseService.getOne(tourId)
     if (result.statusCode === 404) {
-      if (localTour) {
+      if (localTour && localTour.status !== TourStatusType.CREATED) {
         await this.localDatabaseService.deleteTour(tourId)
         return result
       }
@@ -108,15 +105,18 @@ export default class ToursService extends APIService {
     return result
   }
 
-  private async handleTourAddResult (result: SingleApiResponse<Tour>): Promise<SingleApiResponse<Tour>> {
-    if (result.statusCode === 201) {
+  private async handleTourAddResult (result: SingleApiResponse<Tour>, tour: UpdateOrCreateTour): Promise<SingleApiResponse<Tour>> {
+    if (isOfflineResultMessage(result.statusCode, result.statusMessage)) {
+      // save the tour in the local database to not lose the changes if the user refreshes the page
+      await this.localDatabaseService.create(tour)
+    } else if (result.statusCode === 201) {
       await this.localDatabaseService.putTour(result.content!)
     }
     return result
   }
 
   private async handleTourListResult (result: ArrayApiResponse<Tour>): Promise<ArrayApiResponse<Tour>> {
-    if (ToursService.isOffline(result.statusCode, result.statusMessage)) {
+    if (isOfflineResultMessage(result.statusCode, result.statusMessage)) {
       return result
     } else {
       if (!result.content) {
